Fix deprecated importlib.resources.paths
I don't like this fix but `paths` has been deprecated in newer Pythons:

https://importlib-resources.readthedocs.io/en/latest/using.html#migrating-from-legacy

Very annoyingly, we are meant to use `files()` which returns a `MultiplexedPath` instance which is a frustratingly useless class, not least because calling `str()` on it returns `"MultiplexedPath('<path>')"` which is infuriatingly unhelpful. The only way it appears to get a string out of it is to `joinpath()` some fake value and substring it. I don't like this so if you can come up with a better fix please do.